Add UDP and TCP versions with compatible behavior
