+++ exercise_03/udp/server_concurrent.go
@@ -4,11 +4,13 @@ package main
 import (
 	"fmt"
 	"net"
+	"sync"
 )
 
 const requests = 10000
 
-func handleHttpRequest(conn *net.UDPConn, addr *net.UDPAddr, message string) {
+func handleHttpRequest(wg *sync.WaitGroup, conn *net.UDPConn, addr *net.UDPAddr, message string) {
+	defer wg.Done()
 	response := []byte(message)
 	_, err := conn.WriteToUDP(response, addr)
 	if err != nil {
@@ -17,9 +19,7 @@ func handleHttpRequest(conn *net.UDPConn, addr *net.UDPAddr, message string) {
 }
 
 func main() {
-	fmt.Println("Starting UDP server!")
-
-	serverAddr, err := net.ResolveUDPAddr("udp", "localhost:8081")
+	serverAddr, err := net.ResolveUDPAddr("udp", "localhost:8082")
 	if err != nil {
 		fmt.Println("Error resolving address:", err)
 		return
@@ -32,20 +32,18 @@ func main() {
 		return
 	}
 
-	fmt.Println("Waiting for requests...")
-
-	for i := 0; ; i++ {
+	wg := sync.WaitGroup{}
+	for i := 0; i < requests; i++ {
+		wg.Add(1)
 		response := "HTTP/1.1 200 OK"
 		buffer := make([]byte, 1024)
 
-		n, addr, err := conn.ReadFromUDP(buffer)
+		_, addr, err := conn.ReadFromUDP(buffer)
 		if err != nil {
 			response = "HTTP/1.1 500 Internal Server Error"
 		}
 
-		handleHttpRequest(conn, addr, response)
-		
-		fmt.Printf("Request %d from address %s:\n%s\n\n", i+1, addr, string(buffer[:n]))
+		go handleHttpRequest(&wg, conn, addr, response)
 	}
-	fmt.Println("here\n")
+	wg.Wait()
 }
